Tuto text format modified! The text opcode = "TEXT(text)" instead of just "text". This allow multi line texts. Captured texts in textHighlighter are not highlighted. Import/export enabled for tutos. Bug fix: the modified signal was sent by the tuto editor even there was no modifications.
git-svn-id: https://svn.code.sf.net/p/makoureactor/code/trunk@234 916e10b9-6b61-4450-83b7-7c74cc09f2a3

+++ core/field/Section1File.cpp
@@ -159,7 +159,7 @@ bool Section1File::open(const QByteArray &data)
 		memcpy(&posDeb, &constData[posTexts+2], 2);
 		nbTextes = posDeb/2 - 1;
 
-		for(quint16 i=1 ; i<nbTextes ; ++i)
+		for(quint32 i=1 ; i<nbTextes ; ++i)
 		{
 			memcpy(&posFin, &constData[posTexts+2+i*2], 2);
